Adds Settings.
